# packages/api-contracts/openapi/openapi.v1.yaml
openapi: 3.1.0
info:
  title: Opener NetDoor Public and Admin API
  version: 1.11.0
servers:
  - url: https://api.example.com
paths:
...
  /v1/auth/session:
    get:
      summary: Resolve current admin cookie session
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Session state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSessionInfo'
  /v1/auth/logout:
    post:
      summary: Invalidate admin session cookie
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout result
          content:
            application/json:
              schema:
                type: object
                required: [logged_out]
                properties:
                  logged_out:
                    type: boolean
  /v1/admin/tenants:
    get:
      summary: List tenants
      security:
        - bearerAuth: [admin:read]
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Status'
      responses:
        '200':
          description: Tenant list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantListResponse'
    post:
      summary: Create tenant
      security:
        - bearerAuth: [admin:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
  /v1/admin/users:
    get:
      summary: List users in tenant
      security:
        - bearerAuth: [admin:read]
      parameters:
        - in: query
          name: tenant_id
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Status'
      responses:
        '200':
          description: User list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
    post:
      summary: Create user
      security:
        - bearerAuth: [admin:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    delete:
      summary: Delete user
      security:
        - bearerAuth: [admin:write]
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: tenant_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                required: [deleted, id, tenant_id]
                properties:
                  deleted:
                    type: boolean
                  id:
                    type: string
                    format: uuid
                  tenant_id:
                    type: string
                    format: uuid
  /v1/admin/users/block:
    post:
      summary: Block user
      security:
        - bearerAuth: [admin:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLifecycleRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /v1/admin/users/unblock:
    post:
      summary: Unblock user
      security:
        - bearerAuth: [admin:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserLifecycleRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
  /v1/admin/access-keys:
    get:
      summary: List access keys
      security:
        - bearerAuth: [admin:read]
      parameters:
        - in: query
          name: tenant_id
          schema:
            type: string
            format: uuid
        - in: query
          name: user_id
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Status'
      responses:
        '200':
          description: Access key list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessKeyListResponse'
    post:
      summary: Create access key
      security:
        - bearerAuth: [admin:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccessKeyRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessKey'
    delete:
      summary: Revoke access key
      security:
        - bearerAuth: [admin:write]
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: tenant_id
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Revoked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessKey'
  /v1/admin/policies/tenants:
    get:
      summary: List tenant policies or get one by tenant_id
      security:
        - bearerAuth: [admin:read]
      parameters:
        - in: query
          name: tenant_id
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Tenant policy list or single item
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/TenantPolicyListResponse'
                  - $ref: '#/components/schemas/TenantPolicy'
    put:
      summary: Create or update tenant policy
      security:
        - bearerAuth: [admin:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetTenantPolicyRequest'
      responses:
        '200':
          description: Upserted tenant policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantPolicy'
  /v1/admin/policies/users:
    get:
      summary: List user policy overrides or get one by tenant_id+user_id
      security:
        - bearerAuth: [admin:read]
      parameters:
        - in: query
          name: tenant_id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: user_id
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: User policy list or single item
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/UserPolicyOverrideListResponse'
                  - $ref: '#/components/schemas/UserPolicyOverride'
    put:
      summary: Create or update user policy override
      security:
        - bearerAuth: [admin:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetUserPolicyOverrideRequest'
      responses:
        '200':
          description: Upserted user policy override
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPolicyOverride'
  /v1/admin/policies/effective:
    get:
      summary: Get effective policy for tenant/user
      security:
        - bearerAuth: [admin:read]
      parameters:
        - in: query
          name: tenant_id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: user_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Effective policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EffectivePolicy'
  /v1/admin/devices/register:
    post:
      summary: Register or refresh a device with policy enforcement
      security:
        - bearerAuth: [admin:write]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceRequest'
      responses:
        '201':
          description: Device registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
...
  /v1/admin/nodes/detail:
    get:
      summary: Get node details by tenant and id
      security:
        - bearerAuth: [admin:read]
      parameters:
        - in: query
          name: tenant_id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: node_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Node details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Node'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      in: query
      name: offset
      schema:
        type: integer
        minimum: 0
        default: 0
    Status:
      in: query
      name: status
      schema:
        type: string
  schemas:
    Tenant:
      type: object
      required: [id, name, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [active, suspended]
        created_at:
          type: string
          format: date-time
    TenantListResponse:
      type: object
      required: [items, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Tenant'
        limit:
          type: integer
        offset:
          type: integer
    CreateTenantRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
...
    AdminSessionInfo:
      type: object
      required: [authenticated, subject, scopes, expires_at]
      properties:
        authenticated:
          type: boolean
        subject:
          type: string
        tenant_id:
          type: string
          format: uuid
        scopes:
          type: array
          items:
            type: string
        expires_at:
          type: string
          format: date-time
...